[Backtracer][DWARF] Fix SLEB128 sign extension code.
If we read a full 9 bytes of data, the sign extension code causes a
numerical trap because it's trying to negate `Int64.min`.

Fix the sign extension code to shift `Int64(-1)` instead of shifting `1`
and trying to negate afterwards.

rdar://184806813

+++ stdlib/public/RuntimeModule/Dwarf.swift
@@ -340,7 +340,7 @@ extension ImageSource {
     }
 
     if shift < 64 && sign != 0 {
-      value |= -(1 << shift)
+      value |= Int64(-1) << shift
     }
 
     return (addr, value)
